requirement.txt
bleach==2.1.1
bootstrap-admin==0.3.7.1
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.12
django-allauth==0.33.0
django-appconf==1.0.2
django-ckeditor==5.7.1
django-crispy-forms==1.6.1
django-haystack==2.6.1
django-imagekit==4.0.1
django-js-asset==1.2.2
django-mdeditor==0.1.14
django-redis==4.9.0
djangorestframework==3.8.2
emoji==0.4.5
gunicorn==19.7.1
html5lib==1.0b10
idna==2.6
jieba==0.39
Markdown==2.6.8
oauthlib==2.0.3
olefile==0.44
pilkit==2.0
Pillow==4.2.1
Pygments==2.2.0
PyMySQL==0.7.11
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
user-agent==0.1.9
webencodings==0.5.1
Whoosh==2.7.4
